ref: SpotifyAuthState to SpotifyAPI
