+++ MusicWith/Util/Spotify/SpotifySearch.swift
@@ -26,7 +26,7 @@ class SpotifySearch {
         repeat {
             let url = "https://api.spotify.com/v1/search?q=\(query)&type=track&offset=\(_trackStorage.count)&limit=\(CHUNK_SIZE)"
 
-            guard let json   = await getSpotifyJson(url),
+            guard let json   = await SpotifyAPI.shared.getSpotifyAPIJson(url),
                   let tracks = json  [ "tracks" ] as?  [String: Any],
                   let items  = tracks[ "items"  ] as? [[String: Any]] else {
                 return []
